# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NeuroDyne Corp API
  description: |
    REST API for the NeuroDyne Corp project management platform.
    Handles authentication, projects, specifications, tasks, invoices,
    messaging, notifications, file uploads, questionnaires, and contact forms.
  version: 1.0.0
  contact:
    name: NeuroDyne Corp
    email: noreply@neurodynecorp.com
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, project_manager, developer, qa, client]
        phone:
          type: string
        company:
          type: string
        isActive:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, project_manager, developer, qa, client]
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    patch:
      tags: [Users]
      summary: Update user (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                role:
                  type: string
                  enum: [admin, project_manager, developer, qa, client]
                phone:
                  type: string
                company:
                  type: string
                isActive:
                  type: boolean
      responses:
        "200":
          description: User updated
        "404":
          description: User not found
    delete:
      tags: [Users]
      summary: Deactivate user (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: User deactivated

  # ── Projects ──────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications for current user
      security:
        - BearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark notification as read
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Marked as read
...